+++ src/interfaces/options/IKeyVectorOptions.ts
@@ -1,5 +1,5 @@
 import Point from '../../objects/Point'
-import HotKey from '../../utils/HotKey'
+import HotKey from '../../utils/keys/HotKey'
 import { PointCallback } from '../../utils/types'
 
 export default interface IKeyVectorOptions {